Decode Pause and Print Scr correctly.
Tapping "Pause" generates E1,14,77,E1,F0,14,F0,77

We now decode this as:

* PauseBreak/Down
* NumLock/Down
* PauseBreak/Up
* NumLock/Up

Tapping "Print Scr" generates E0,12,E0,7C. This becomes:

We now decode this as:

* PrintScreen/SingleShot
* PrintScreen2/SingleShot

Note the addition of the "SingleShot" KeyState, for keys that don't
have a Make/Break pair. This avoids users waiting a "Break" that never
happens.